Enable conversion of Pubkeys to Pedersen Commitments
This is the reverse of 36b36581ccce9071276464ad10cbc65e229257fe.
It becomes particularly helpful for doing intermediary sums
of pedersen commitments (via hacky use of ec_pubkey_combine()).

Note: This is probably not a great idea and should /not/ be included
upstream; it's just useful for researching some new protocols

Signed-off-by: Sam Stuewe <stuewe@mit.edu>